requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
priority==2.0.0
psycopg==3.1.17
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pytest==7.4.4
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.25.0
wsproto==1.2.0
